[INJI-691] move backup & restore related components to folder
