+++ machines/backupAndRestore/backup.typegen.ts
@@ -48,9 +48,7 @@ export interface Typegen0 {
       | 'error.platform.backup.backingUp.uploadBackupFile:invocation[0]'
       | 'error.platform.backup.backingUp.zipBackupFile:invocation[0]';
     sendDataBackupStartEvent: 'DATA_BACKUP' | 'FETCH_DATA';
-    sendDataBackupSuccessEvent:
-      | 'done.invoke.backup.backingUp.uploadBackupFile:invocation[0]'
-      | 'done.invoke.backup.backingUp.zipBackupFile:invocation[0]';
+    sendDataBackupSuccessEvent: 'done.invoke.backup.backingUp.uploadBackupFile:invocation[0]';
     setDataFromStorage: 'STORE_RESPONSE';
     setFileName: 'FILE_NAME';
   };
@@ -60,7 +58,7 @@ export interface Typegen0 {
   };
   eventsCausingServices: {
     checkStorageAvailability: 'DATA_BACKUP' | 'FETCH_DATA';
-    uploadBackupFile: never;
+    uploadBackupFile: 'done.invoke.backup.backingUp.zipBackupFile:invocation[0]';
     writeDataToFile: 'STORE_RESPONSE';
     zipBackupFile: 'FILE_NAME';
   };
